Move GANLoop to experimental
- This isn't really coherent with the current interface of TrainingLoop.
  It should be. But it will need refactoring. This is not currently a
  priority.

Clean up GAN
